# Akka.Cluster.Sharding performance benchmarks