--- a/src/benchmark/Akka.Cluster.Benchmarks/Akka.Cluster.Benchmarks.csproj
+++ b/src/benchmark/Akka.Cluster.Benchmarks/Akka.Cluster.Benchmarks.csproj
@@ -5,4 +5,19 @@
         <TargetFramework>$(NetCoreTestVersion)</TargetFramework>
     </PropertyGroup>
 
+    <ItemGroup>
+        <PackageReference Include="BenchmarkDotNet" Version="$(BenchmarkDotNetVersion)" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\..\contrib\cluster\Akka.Cluster.Sharding\Akka.Cluster.Sharding.csproj" />
+      <ProjectReference Include="..\..\contrib\persistence\Akka.Persistence.Sqlite\Akka.Persistence.Sqlite.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <Compile Include="..\Akka.Benchmarks\Configurations\Configs.cs">
+        <Link>Configs.cs</Link>
+      </Compile>
+    </ItemGroup>
+
 </Project>