--- a/src/benchmark/Akka.Cluster.Benchmarks/Program.cs
+++ b/src/benchmark/Akka.Cluster.Benchmarks/Program.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using System.Reflection;
+using BenchmarkDotNet.Running;
 
 namespace Akka.Cluster.Benchmarks
 {
@@ -6,7 +8,7 @@ namespace Akka.Cluster.Benchmarks
     {
         static void Main(string[] args)
         {
-            Console.WriteLine("Hello World!");
+            BenchmarkSwitcher.FromAssembly(Assembly.GetExecutingAssembly()).Run(args);
         }
     }
 }