@Aaronontheweb 2021-08-17
The numbers come out to 2,400 per second - exactly what was measured on https://github.com/akkadotnet/akka.net/issues/5203