@Aaronontheweb 2021-08-18
Spawn benchmarks:

``` ini

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19041.1165 (2004/May2020Update/20H1)
AMD Ryzen 7 1700, 1 CPU, 16 logical and 8 physical cores
.NET SDK=5.0.302
  [Host]     : .NET Core 3.1.17 (CoreCLR 4.700.21.31506, CoreFX 4.700.21.31502), X64 RyuJIT
  Job-BWMBDZ : .NET Core 3.1.17 (CoreCLR 4.700.21.31506, CoreFX 4.700.21.31502), X64 RyuJIT

InvocationCount=1  IterationCount=1  LaunchCount=5  
RunStrategy=ColdStart  UnrollFactor=1  WarmupCount=0  

```
|        Method |   StateMode | EntityCount | RememberEntities |     Mean |    Error |   StdDev |   Median |
|-------------- |------------ |------------ |----------------- |---------:|---------:|---------:|---------:|
| **SpawnEntities** | **Persistence** |        **1000** |            **False** |  **1.836 s** | **2.1365 s** | **0.5548 s** |  **1.446 s** |
| **SpawnEntities** | **Persistence** |        **1000** |             **True** |  **2.100 s** | **1.5785 s** | **0.4099 s** |  **1.901 s** |
| **SpawnEntities** |       **DData** |        **1000** |            **False** |  **1.033 s** | **0.0969 s** | **0.0252 s** |  **1.033 s** |
| **SpawnEntities** |       **DData** |        **1000** |             **True** | **74.788 s** | **9.7173 s** | **2.5235 s** | **74.670 s** |